move functionality of script into library
